+++ spec/q-spec.js
@@ -2343,6 +2343,15 @@ describe("node support", function () {
             });
         });
 
+        it("allows bound function to support missing arguments", function() {
+            return Q.nfbind(function (a, b, c, d, callback) {
+                callback(null, !d);
+            }).call({},1,2,3)
+            .then(function(nod) {
+                expect(nod).toBe(true);
+            });
+        });
+
     });
 
     describe("nbind", function () {